project_underhill/core/database.py
import sqlalchemy
from sqlalchemy import MetaData, create_engine
import databases

from dotenv import load_dotenv
from os import getenv

load_dotenv()

DATABASE_URL = getenv("DATABASE_URL")

database = databases.Database(DATABASE_URL)

metadata = MetaData()

engine = create_engine(DATABASE_URL)
