pyproject.toml
[tool.poetry]
name = "project_underhill"
version = "0.1.0"
description = ""
authors = ["Erik Seguinte <ErikSeguinte@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
fastapi = {extras = ["all"], version = "^0.58.0"}
databases = {extras = ["postgresql"], version = "^0.3.2"}
gunicorn = "^20.0.4"
passlib = {extras = ["argon2"], version = "^1.7.2"}
sqlalchemy = "^1.3.18"
alembic = "^1.4.2"
python-dotenv = "^0.14.0"
psycopg2-binary = "^2.8.5"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
pip = "^20.1.1"
setuptools = "^47.3.1"
pre-commit = "^2.5.1"
black = "^19.10b0"
lorem-text = "^1.5"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
